fix(theme): Add darkGray color resource
- Added a new color resource, `darkGray` (#A9A9A9), to the `colors.xml` file.

This change fixes a critical `Unresolved reference` build error that was occurring in the `ProductAdapter`. The adapter was attempting to use `R.color.darkGray` for the tint of the bordered wishlist icon when an item was not in the user's wishlist, but the color had not yet been defined.

By adding this color resource, the build error is resolved and the wishlist icon's default (unselected) state will now render correctly.